requirements.txt
# Web Framework
python-fasthtml>=0.3.0
uvicorn>=0.25.0

# AI/LLM
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0

# File Processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pillow>=10.1.0
python-magic>=0.4.27

# YouTube & Media
yt-dlp>=2023.12.0
pytube>=15.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
